+++ Source/PlayScene/UI/MiniChara/MiniCharaGlocken.cpp
@@ -27,7 +27,7 @@ void MiniCharaGlocken::Update(MiniChara& _self)
 	{
 		float angle{ DirectX::XM_2PI * (_self.animTimeLeft_ / _self.config_.totalAnimTime) + offsetAngle };
 		Vector2Int size{ _self.imageSize_ };
-		size.x += std::sinf(angle) * 10.0f;
+		size.x += static_cast<int>(std::sinf(angle) * 10.0f);
 		config.scale(size);
 		CONTEXT.SetRefLayout(&config);
 		CONTEXT.DrawImage(_self.hImage_);
