+++ Source/PlayScene/UI/MiniChara/MiniCharaMonkitty.cpp
@@ -24,7 +24,7 @@ void MiniCharaMonkitty::Update(MiniChara& _self)
 
 	float angle{ DirectX::XM_2PI * (_self.animTimeLeft_ / _self.config_.totalAnimTime) };
 	Vector2Int size{ _self.imageSize_ };
-	size.x += std::sinf(angle) * 10.0f;
+	size.x += static_cast<int>(std::sinf(angle) * 10.0f);
 	config.scale(size);
 	CONTEXT.SetRefLayout(&config);
 	CONTEXT.DrawImage(_self.hImage_, angle);
